Move files that were in nested directories to the root of model dir now that most of the extra files have been removed
